Fixed log variable format specifier.

+++ src/cwebsocket/server.c
@@ -454,12 +454,12 @@ int cwebsocket_server_read_data(cwebsocket_connection *connection) {
 
 		size_t utf8_code_points = 0;
 		if(utf8_count_code_points((uint8_t *)payload, &utf8_code_points)) {
-			syslog(LOG_ERR, "cwebsocket_server_read_data: received %lld byte malformed utf-8 text payload: %s\n", payload_length, payload);
+			syslog(LOG_ERR, "cwebsocket_server_read_data: received %llu byte malformed utf-8 text payload: %s\n", payload_length, payload);
 			cwebsocket_server_onerror(connection, "received malformed utf-8 payload");
 			return -1;
 		}
 
-		syslog(LOG_DEBUG, "cwebsocket_server_read_data: received %lld byte text payload: %s", payload_length, payload);
+		syslog(LOG_DEBUG, "cwebsocket_server_read_data: received %llu byte text payload: %s", payload_length, payload);
 
 		if(connection->subprotocol != NULL && connection->subprotocol->onmessage != NULL) {
 
@@ -570,7 +570,7 @@ int cwebsocket_server_read_data(cwebsocket_connection *connection) {
 		for(i=0; i<payload_length; i++) {
 			reason[i] = reason[i] ^ frame.masking_key[i%4];
 		}
-		syslog(LOG_DEBUG, "cwebsocket_server_read_data: received CLOSE control frame. payload_length=%lld, code=%i, reason=%s", payload_length, code, reason);
+		syslog(LOG_DEBUG, "cwebsocket_server_read_data: received CLOSE control frame. payload_length=%llu, code=%i, reason=%s", payload_length, code, reason);
 		cwebsocket_server_close_connection(connection, (uint32_t)code, (const char *) reason);
 		return 0;
 	}
@@ -630,8 +630,8 @@ ssize_t cwebsocket_server_write_data(cwebsocket_connection *connection, const ch
 		return -1;
 	}
 
-	syslog(LOG_DEBUG, "cwebsocket_server_write_data: bytes_written=%zu, frame_length=%i, opcode=%#04x, payload_len=%lld, payload=%s\n",
-			bytes_written, frame_length, code, (long long)payload_len, data);
+	syslog(LOG_DEBUG, "cwebsocket_server_write_data: bytes_written=%zu, frame_length=%i, opcode=%#04x, payload_len=%llu, payload=%s\n",
+			bytes_written, frame_length, code, payload_len, data);
 
 	return bytes_written;
 }
